+++ internal/interfaces/sqlite/db.go
@@ -13,7 +13,7 @@ var tablesStmts []string = []string{
 	`CREATE TABLE IF NOT EXISTS job(name TEXT PRIMARY KEY, command TEXT, tick TEXT)`,
 	`CREATE TABLE IF NOT EXISTS execution(
 		job_name TEXT, date DATE, STATUS TEXT, 
-		mem_usage REAL, cpu_usage REAL, net_in INTEGER, net_out INTEGER
+		mem_usage INTEGER, cpu_usage REAL, net_in INTEGER, net_out INTEGER
 	)`,
 }
